refactor directories, provisioning/tinyusb
